move setup controller to core/controller
